Refactor out of BaseMap
